Enum libcoinche::bid::BidError

pub enum BidError {
    AuctionClosed,
    TurnError,
    NonRaisedTarget,
    AuctionRunning,
    NoContract,
    OverCoinche,
}

Possible error occuring during an Auction.

Variants

AuctionClosed

The auction was closed and does not accept more contracts.

TurnError

A player tried bidding before his turn.

NonRaisedTarget

The given bid was not higher than the previous one.

AuctionRunning

Cannot complete the auction when it is still running.

NoContract

No contract was offered during the auction, it cannot complete.

OverCoinche

The contract was coinched too many times.