Enum libcoinche::bid::Target

pub enum Target {
    Contract80,
    Contract90,
    Contract100,
    Contract110,
    Contract120,
    Contract130,
    Contract140,
    Contract150,
    Contract160,
    ContractCapot,
}

Goal set by a contract.

Determines the winning conditions and the score on success.

Variants

Contract80

Team must get 80 points

Contract90

Team must get 90 points

Contract100

Team must get 100 points

Contract110

Team must get 110 points

Contract120

Team must get 120 points

Contract130

Team must get 130 points

Contract140

Team must get 140 points

Contract150

Team must get 150 points

Contract160

Team must get 160 points

ContractCapot

Team must win all tricks

Methods

impl Target

fn score(&self) -> i32

Returns the score this target would give on success.

fn to_str(&self) -> &'static str

fn victory(&self, points: i32, capot: bool) -> bool

Determines whether this target was reached.